Move `ClientMapper` and `ServerEntityMap` to `client_mapper` submodule
Because of internal state of `ServerEntityMap` that should be private.
